show option A and B; will drop B after this and start finishing off A, currently incomplete
