+++ package.json
@@ -120,7 +120,10 @@
     "@elastic/elasticsearch": "^7.10.0",
     "@hapi/hapi": "^20.1.2",
     "@koa/router": "^9.0.1",
+    "@types/aws-lambda": "^8.10.77",
+    "@types/connect": "^3.4.34",
     "@types/node": "^13.7.4",
+    "@types/pino": "^6.3.8",
     "ajv": "^6.12.6",
     "apollo-server-express": "^2.10.1",
     "aws-sdk": "^2.622.0",
